+++ tests/jsonpickle_test.py
@@ -103,6 +103,59 @@ class PicklingTestCase(unittest.TestCase):
         self.pickler.reset()
         self.unpickler.reset()
 
+    @unittest.skipIf(not PY2, 'Python 2-specific base85 test')
+    def test_base85_always_false_on_py2(self):
+        pickler = jsonpickle.pickler.Pickler(use_base85=True)
+        self.assertFalse(pickler.use_base85)
+
+    @unittest.skipIf(PY2, 'Base85 not supported on Python 2')
+    def test_base85_default_py3(self):
+        """Ensure Python 2 allows use_base85 as default on Python 3"""
+        pickler = jsonpickle.pickler.Pickler()
+        self.assertTrue(pickler.use_base85)
+
+    @unittest.skipIf(PY2, 'Base85 not supported on Python 2')
+    def test_base85_override_py3(self):
+        """Ensure the Python 2 check still lets us set use_base85 on Python 3"""
+        pickler = jsonpickle.pickler.Pickler(use_base85=False)
+        self.assertFalse(pickler.use_base85)
+
+    @unittest.skipIf(PY2, 'Base85 not supported on Python 2')
+    def test_bytes_default_base85(self):
+        data = os.urandom(16)
+        encoded = util.b85encode(data)
+        self.assertEqual({tags.B85: encoded}, self.pickler.flatten(data))
+
+    @unittest.skipIf(PY2, 'Base85 not supported on Python 2')
+    def test_py3_bytes_base64_override(self):
+        pickler = jsonpickle.pickler.Pickler(use_base85=False)
+        data = os.urandom(16)
+        encoded = util.b64encode(data)
+        self.assertEqual({tags.B64: encoded}, pickler.flatten(data))
+
+    @unittest.skipIf(not PY2, 'Python 2-specific base64 test')
+    def test_py2_default_base64(self):
+        data = os.urandom(16)
+        encoded = util.b64encode(data)
+        self.assertEqual({tags.B64: encoded}, self.pickler.flatten(data))
+
+    @unittest.skipIf(PY2, 'Base85 not supported on Python 2')
+    def test_decode_base85(self):
+        pickled = {tags.B85: 'P{Y4;Xv4O{u^=-c'}
+        expected = u'P\u00ffth\u00f6\u00f1 3!'.encode('utf-8')
+        self.assertEqual(expected, self.unpickler.restore(pickled))
+
+    @unittest.skipIf(PY2, 'Base85 not supported on Python 2')
+    def test_base85_still_handles_base64(self):
+        pickled = {tags.B64: 'UMO/dGjDtsOxIDMh'}
+        expected = u'P\u00ffth\u00f6\u00f1 3!'.encode('utf-8')
+        self.assertEqual(expected, self.unpickler.restore(pickled))
+
+    @unittest.skipIf(not PY2, 'Python 2-specific base85 test')
+    def test_base85_crashes_py2(self):
+        with self.assertRaises(NotImplementedError):
+            self.unpickler.restore({tags.B85: 'P{Y4;Xv4O{u^=-c'})
+
     def test_string(self):
         self.assertEqual('a string', self.pickler.flatten('a string'))
         self.assertEqual('a string', self.unpickler.restore('a string'))